Extract out String formatting for SRP
Create the Formatter class to handle String formatting instead of TextUI
